refactor(cli): carve the CLI app behind the sipi_cli_main FFI entry
Move the CLI11 app and every subcommand body out of the binary's main
translation unit into a new //src/cli:cli_app cc_library, defining the
extern "C" sipi_cli_main entry the seam already declares (ffi/sipi_ffi.h).
src/cli/sipi.cpp becomes a thin main that just calls sipi_cli_main.

This is what lets the Rust HTTP shell (Phase C) drive the offline
subcommands: it can link cli_app and call sipi_cli_main without colliding
with Rust's own main, and Rust owns process teardown after dispatch.

The subcommand callbacks no longer call exit() — a process exit from inside
the FFI would skip the Rust caller's cleanup (OTel flush, Sentry close).
Each leaf callback records its result into a sipi_exit_code captured by
reference, returned after CLI11_PARSE. require_subcommand(1) guarantees
exactly one leaf assigns it; the convert/verify parent callbacks already
self-skip when a nested subcommand matched, so the assignment is
single-shot regardless of callback order. The only remaining abort() is the
my_terminate_handler crash path (intended Sentry inproc reporting).

Behaviour unchanged: //src/cli:sipi builds and the cli + cli_json e2e tests
pass (--version → 0, missing subcommand → 106, missing file → 105). Update
the four config-surface doc references + the CLAUDE.md component table from
sipi.cpp to cli_app.cpp (the rename's blast radius); docs/archive/ stays as
the point-in-time record.
